posts commit

+++ controllers/controllerPosts.js
@@ -1,9 +1,10 @@
 import postModel from "../models/postModel.js";
+import userModel from "../models/userModel.js";
 
 const getPostsByPage = async (page, isCreate) => {
   const currentPage = Number(page);
   const LIMIT =
-  isCreate & (currentPage === 1) ? 8 : isCreate & (currentPage > 1) ? 7 : 8;
+    isCreate & (currentPage === 1) ? 8 : isCreate & (currentPage > 1) ? 7 : 8;
   const startIndex = (currentPage - 1) * LIMIT;
 
   try {
@@ -24,7 +25,10 @@ export const getPosts = async (req, res) => {
   const isCreate = false;
   const { page } = req.query;
   try {
-    const { posts, currentPage, totalPages } = await getPostsByPage( page, isCreate );
+    const { posts, currentPage, totalPages } = await getPostsByPage(
+      page,
+      isCreate
+    );
     res.status(200).json({ posts, currentPage, totalPages });
   } catch (error) {
     res.status(404).json({ message: error.message });
@@ -72,8 +76,16 @@ export const createPost = async (req, res) => {
   const isCreate = true;
   const { page } = req.query;
   const post = req.body;
-  const newPost = { ...post, creator: req.userId };
   try {
+    const postCreator = await userModel.findById(req.userId);
+    const newPost = {
+      ...post,
+      creatorId: req.userId,
+      isCreatorAdmin: postCreator.isAdmin,
+      creatorLevel: postCreator.level,
+      isCreatorOwner: postCreator.isOwner,
+    };
+
     const createdPost = await new postModel(newPost).save();
     const { posts, currentPage, totalPages } = await getPostsByPage(
       page,
@@ -88,8 +100,32 @@ export const createPost = async (req, res) => {
 
 export const deletePost = async (req, res) => {
   try {
-    await postModel.findByIdAndRemove(req.params.id);
-    res.status(200).json("Deleted Successfully");
+    const sender = await userModel.findById(req.userId);
+    const post = await postModel.findById(req.params.id);
+    if (
+      (sender.isOwner && !post.isCreatorOwner) ||
+      String(sender._id) === post.creatorId
+    ) {
+      await postModel.findByIdAndRemove(req.params.id);
+      res.status(200).json("Deleted Successfully");
+    } else if (
+      sender.isOwner &&
+      post.isCreatorOwner &&
+      (String(sender._id) === post.creatorId || sender.level > post.creatorLevel)
+    ) {
+      await postModel.findByIdAndRemove(req.params.id);
+      res.status(200).json("Deleted Successfully");
+    } else if (
+      !sender.isOwner &&
+      !post.isCreatorOwner &&
+      sender.isAdmin &&
+      sender.level > post.creatorLevel
+    ) {
+      await postModel.findByIdAndRemove(req.params.id);
+      res.status(200).json("Deleted Successfully");
+    } else {
+      res.status(403).json("Unauthorized to delete post");
+    }
   } catch (error) {
     res.status(404).json({ message: error.message });
   }
@@ -97,12 +133,17 @@ export const deletePost = async (req, res) => {
 
 export const updatePost = async (req, res) => {
   try {
-    const updatedPost = await postModel.findByIdAndUpdate(
-      req.params.id,
-      { $set: req.body },
-      { new: true }
-    );
-    res.status(200).json(updatedPost);
+    const post = await postModel.findById(req.params.id);
+    if (post.creatorId === req.userId) {
+      const updatedPost = await postModel.findByIdAndUpdate(
+        req.params.id,
+        { $set: req.body },
+        { new: true }
+      );
+      res.status(200).json(updatedPost);
+    } else {
+      res.status(403).json("Unauthorized to update post");
+    }
   } catch (error) {
     res.status(404).json({ message: error.message });
   }
@@ -120,7 +161,9 @@ export const likePost = async (req, res) => {
       post.likes.splice(index, 1);
       // post.likes = post.likes.filter((id) => id !== String(req.userId));
     }
-    const likedPost = await postModel.findByIdAndUpdate(req.params.id, post, { new: true });
+    const likedPost = await postModel.findByIdAndUpdate(req.params.id, post, {
+      new: true,
+    });
     res.status(200).json(likedPost);
   } catch (error) {
     res.status(404).json({ message: error.message });
@@ -129,14 +172,24 @@ export const likePost = async (req, res) => {
 
 export const commentPost = async (req, res) => {
   const { id } = req.params;
-  const { value } = req.body;
-  const comment = { ...value, creator: req.userId };
+  const { comment } = req.body;
 
   try {
+    const commentCreator = await userModel.findById(req.userId);
+    const newComment = {
+      ...comment,
+      creatorId: req.userId,
+      isCreatorAdmin: commentCreator.isAdmin,
+      creatorLevel: commentCreator.level,
+      isCreatorOwner: commentCreator.isOwner,
+    };
+
     const post = await postModel.findById(id);
-    // post.comments = [...post.comments, value] OR
-    post.comments.push(comment);
-    const updatedPost = await postModel.findByIdAndUpdate(id, post, { new: true, });
+    // post.comments = [...post.comments, newComment] OR
+    post.comments.push(newComment);
+    const updatedPost = await postModel.findByIdAndUpdate(id, post, {
+      new: true,
+    });
     res.status(200).json(updatedPost);
   } catch (error) {
     res.status(401).json({ message: error.message });
@@ -144,29 +197,21 @@ export const commentPost = async (req, res) => {
 };
 
 export const deleteComment = async (req, res) => {
-  const cId = req.params.id;
-  const { pId } = req.body;
-  // console.log(typeof req.body)
+  const commentId = req.params.id;
+  const { postId } = req.body;
   try {
-    const foundPost = await postModel.findById(pId);
-    foundPost.comments = foundPost.comments.filter((com) => com.id !== String(cId));
-    const post = await postModel.findByIdAndUpdate(pId, foundPost, { new: true });
+    const foundPost = await postModel.findById(postId);
+    foundPost.comments = foundPost.comments.filter(
+      (com) => com.id !== String(commentId)
+    );
+    const post = await postModel.findByIdAndUpdate(pId, foundPost, {
+      new: true,
+    });
     res.status(200).json(post);
   } catch (error) {
     res.status(404).json({ message: error.message });
   }
 };
-
-
-// the below code is same with the above, just that some logics were added.
-// export const likePost = async (req, res) => {
-//    try {
-//       const likedPost = await postModel.findByIdAndUpdate(req.params.id, {likeCount: likeCount + 1}, { new: true });
-//       res.status(200).json(likedPost);
-//    } catch (error) {
-//       res.status(404).json({message: error.message})
-//    }
-// }
 
 // export const likePost = async (req, res) => {
 //       const likedPost = await postModel.findOneAndUpdate({_id:req.params.id}, {$inc:{ likeCount:1  }}, { new: true });
@@ -175,5 +220,3 @@ export const deleteComment = async (req, res) => {
 //       res.status(404).json({message: error.message})
 //    }
 // }
-
-
